Migrate the home components to align with Phlex conversions
